componentize current topic + clean up
