+++ src/session/discussion/Discussion.js
@@ -1,34 +1,9 @@
 import React from 'react'
+import Current from './Current'
 
 export default function Discussion(props) {
 
   const [activeTab, setActiveTab] = React.useState("CURRENT")
-  const [timerString, setTimerString] = React.useState("")
-
-  React.useEffect(() => timer())
-
-  let timer = () => {
-    setInterval(() => {
-      if (props.topics.currentDiscussionItem !== undefined && props.topics.currentDiscussionItem.endTime !== undefined) {
-        const timestamp = Date.parse(props.topics.currentDiscussionItem.endTime)
-        let secondsLeft = Math.max(0, Math.round((timestamp - Date.now()) / 1000))
-        let result = getTimerStringFromSecondsLeft(secondsLeft)
-        if (timerString !== result) {
-          setTimerString(result)
-        }
-      }
-    }, 500);
-  }
-
-  let getTimerStringFromSecondsLeft = (secondsLeft) => {
-    let min = Math.floor(secondsLeft / 60)
-    let seconds = Math.floor(secondsLeft % 60)
-    let sec = seconds.toString()
-    if (sec.length < 2) {
-      sec = '0' + sec
-    }
-    return min + ':' + sec;
-  }
 
   let queueButtonStyle = activeTab === "QUEUE"
     ? "shadow-custom"
@@ -41,37 +16,6 @@ export default function Discussion(props) {
   let pastButtonStyle = activeTab === "PAST"
     ? "shadow-custom"
     : ""
-
-  let current = () => {
-    if (props.topics.currentDiscussionItem !== undefined && props.topics.currentDiscussionItem.text !== undefined) {
-      return (
-        <div className="h-full flex flex-col">
-          <div className="mt-2">
-            <h2>Current Discussion Topic</h2>
-          </div>
-          
-
-          <div className="overflow-scroll px-5 mt-2">
-            <h1>{props.topics.currentDiscussionItem.text}</h1>
-          </div>
-
-          <div className="flex flex-grow flex-col justify-end mt-3">
-            <h2>{timerString}</h2>
-          </div>
-        </div>
-      )
-    } else {
-      return <h2>All conversation topics completed!</h2>
-    }
-  }
-
-  let activeTabElement = () => {
-    return activeTab === "QUEUE"
-      ? "QUEUE"
-      : activeTab === "PAST"
-          ? "PAST"
-          : current()
-  }
   
   return (
     <div className="pb-2">
@@ -82,7 +26,11 @@ export default function Discussion(props) {
       </div>
 
       <div className="h-70vh sm:h-80vh m-auto w-85w border rounded">
-        {activeTabElement()}
+        {activeTab === "QUEUE"
+            ? "QUEUE"
+            : activeTab === "PAST"
+              ? "PAST"
+              : <Current topic={props.topics.currentDiscussionItem}/>}
       </div>
     </div>
   )
